Replaced EnterspeedJsonEntity and EnterspeedObjectEntity with EnterspeedSourceEntity
