+++ src/Enterspeed.Source.Sdk/Api/Models/EnterspeedSourceEntity.cs
@@ -1,6 +1,6 @@
 ﻿namespace Enterspeed.Source.Sdk.Api.Models
 {
-    public class EnterspeedObjectEntity : IEnterspeedObjectEntity
+    public class EnterspeedSourceEntity : IEnterspeedSourceEntity
     {
         public string Id { get; set; }
         public string Type { get; set; }
@@ -8,5 +8,12 @@
         public string[] Redirects { get; set; }
         public string ParentId { get; set; }
         public object Properties { get; set; }
+
+        public EnterspeedSourceEntity(string id, string type, object properties = null)
+        {
+            Id = id;
+            Type = type;
+            Properties = properties;
+        }
     }
 }
